# harness/dist/adapter.openapi.json
{
  "components": {
    "schemas": {
      "CleaningRobotCommandRequest": {
        "description": "清洁机器人命令请求。",
        "properties": {
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "robot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Robot Id"
          },
          "scheduled_cleaning_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Cleaning At"
          },
          "scheduling_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduling Method"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "CleaningRobotCommandRequest",
        "type": "object"
      },
      "CleaningRobotCommandResult": {
        "description": "清洁机器人命令结果。",
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "robot_id": {
            "title": "Robot Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "vendor_payload": {
            "additionalProperties": true,
            "title": "Vendor Payload",
            "type": "object"
          }
        },
        "required": [
          "robot_id",
          "command",
          "status"
        ],
        "title": "CleaningRobotCommandResult",
        "type": "object"
      },
...
      "DeadLetterReplayRequest": {
        "description": "POST /api/v1/deadletters/replay 请求(jsonl 缺省=重放当前死信队列)。",
        "properties": {
          "jsonl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jsonl"
          }
        },
        "title": "DeadLetterReplayRequest",
        "type": "object"
      },
      "DeadLetterReplayResult": {
        "description": "死信重放结果(DedupeCache 协同:已投递的跳过,下游仍只见一次)。",
        "properties": {
          "enqueued": {
            "title": "Enqueued",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "enqueued",
          "skipped"
        ],
        "title": "DeadLetterReplayResult",
        "type": "object"
      },
...
      "FlyCartCommandRequest": {
        "description": "FlyCart 命令请求(device_sn 必填由 dispatch 校验)。",
        "properties": {
          "cmd": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmd"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "device_sn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Sn"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "task": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "title": "FlyCartCommandRequest",
        "type": "object"
      },
      "FlyCartCommandResult": {
        "description": "FlyCart 命令结果。",
        "properties": {
          "bid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid"
          },
          "command": {
            "title": "Command",
            "type": "string"
          },
          "device_sn": {
            "title": "Device Sn",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "vendor_payload": {
            "additionalProperties": true,
            "title": "Vendor Payload",
            "type": "object"
          }
        },
        "required": [
          "device_sn",
          "command",
          "status"
        ],
        "title": "FlyCartCommandResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecentEvents": {
        "description": "GET /api/v1/events/recent 响应(newest-first)。",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/UnifiedEventView"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "count",
          "events"
        ],
        "title": "RecentEvents",
        "type": "object"
      },
...
      "UnifiedEventView": {
        "description": "统一事件视图。",
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "severity": {
            "default": "info",
            "title": "Severity",
            "type": "string"
          },
          "sn": {
            "default": "",
            "title": "Sn",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "ts": {
            "default": "",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "source",
          "event_type"
        ],
        "title": "UnifiedEventView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "XingluoCommandRequest": {
        "description": "星逻命令请求(条件必填由 dispatch 校验→400)。",
        "properties": {
          "camera_lens": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Lens"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "mission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "uav_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uav Id"
          }
        },
        "title": "XingluoCommandRequest",
        "type": "object"
      },
      "XingluoCommandResult": {
        "description": "星逻命令结果(succeeded|accepted)。",
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "mission_batch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Batch"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uav_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uav Id"
          },
          "vendor_payload": {
            "additionalProperties": true,
            "title": "Vendor Payload",
            "type": "object"
          }
        },
        "required": [
          "command",
          "status"
        ],
        "title": "XingluoCommandResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Gd-港电 云云对接多平台适配器",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/commands/cleaning-robot": {
      "post": {
        "description": "@brief 清洁机器人命令(ack 受理语义)",
        "operationId": "commands_cleaning_robot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleaningRobotCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleaningRobotCommandResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commands Cleaning Robot"
      }
    },
    "/api/v1/commands/flycart": {
      "post": {
        "description": "@brief FlyCart 命令(bid 终态礼貌轮询)",
        "operationId": "commands_flycart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlyCartCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlyCartCommandResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commands Flycart"
      }
    },
    "/api/v1/commands/xingluo": {
      "post": {
        "description": "@brief 星逻无人机命令(takeoff 成功后登记批次续跟)",
        "operationId": "commands_xingluo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/XingluoCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XingluoCommandResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commands Xingluo"
      }
    },
...
    "/api/v1/deadletters/replay": {
      "post": {
        "description": "@brief 13-R-AD-3:死信重放(已投递的经 DedupeCache 跳过)",
        "operationId": "deadletters_replay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeadLetterReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterReplayResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deadletters Replay"
      }
    },
    "/api/v1/events/recent": {
      "get": {
        "description": "@brief 旁路环形缓冲查询(钳制 1..容量,newest-first,不消费队列)",
        "operationId": "events_recent",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "title": "Event Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentEvents"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events Recent"
      }
    },
...
  }
}